==> code/mention-classification/round4/01-sample_social_group_category_annotation-round3.ipynb <==
import os
import json

import numpy as np
import pandas as pd

from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity


## Load data

data_path = os.path.join('..', '..', '..', 'data')
annotations_path = os.path.join(data_path, 'annotations', 'group_mention_categorization')

### Group mentions with attribute annotations

# TODO: update file after integrating reviewed cases
fp = os.path.join(annotations_path, 'social-group-mention-categorization-llm-review', 'consolidated_annotations.tsv')
labeled_mentions = pd.read_csv(fp, sep='\t')

labeled_mentions['attribute'] = labeled_mentions.q_id.str.removesuffix('_attributes')

attributes = ['economic', 'non-economic']
labeled_mentions = labeled_mentions[labeled_mentions.attribute.isin(attributes)]

# inspect distribution of econ attributes
tmp = labeled_mentions[labeled_mentions.attribute==attributes[0]].value_counts(['category', 'label']).reset_index().pivot(index='category', columns='label', values='count')
tmp.fillna(0, inplace=True)
tmp = tmp.astype(int)
tmp['n'] = tmp[['No', 'Yes']].sum(axis=1)
tmp['prevalence'] = tmp['Yes'] / tmp['n']
tmp.round(3)

# inspect distribution of non-econ attributes
tmp = labeled_mentions[labeled_mentions.attribute==attributes[1]].value_counts(['category', 'label']).reset_index().pivot(index='category', columns='label', values='count')
tmp.fillna(0, inplace=True)
tmp['n'] = tmp[['No', 'Yes']].sum(axis=1)
tmp['prevalence'] = tmp['Yes'] / tmp['n']
tmp.round(3)

**NOTE:** some categories like 'economic > education level' or 'non-economic > religion' are severely underrepresented and hence likely unrepresentative of the conceptual core of the given category. This creates problems for (few-shot) learning. 

&rarr; need to sample more examples

labeled_mentions.query("category=='religion' and label=='Yes'").mention.drop_duplicates().values

### Social group mentions without attribute annotations

fp = os.path.join(data_path, 'labeled', 'manifesto_sentences_predicted_group_mentions_spans.tsv')
unlabeled_mentions = pd.read_csv(fp, sep='\t')

# subset to social group mentions
unlabeled_mentions = unlabeled_mentions.loc[unlabeled_mentions.label=='social group', ['sentence_id', 'sentence_text', 'span_nr', 'text']]

# discard already labeled mentions
unlabeled_mentions['mention_id'] = unlabeled_mentions['sentence_id'] + '-' + unlabeled_mentions['span_nr'].astype(str)
unlabeled_mentions = unlabeled_mentions[~unlabeled_mentions.mention_id.isin(labeled_mentions['mention_id'].tolist())]

# remove duplicates
unlabeled_mentions = unlabeled_mentions.sample(frac=1.0, random_state=42)
unlabeled_mentions = unlabeled_mentions.drop_duplicates('text')

## Sample additional unlabeled mentions for attribute annotation 

apply two criteria:

- similarity to attribute category definition
- try to balance the label distribution by over-sampling currently under-represented attribute categories

### Determine attribute category-specific "quotas"

ANNOTATION_BUDGET = 200

# inspect distribution of non-econ attributes
tmp = labeled_mentions.value_counts(['category', 'label']).reset_index().pivot(index='category', columns='label', values='count')
tmp.fillna(0, inplace=True)
tmp = tmp.astype(int)
tmp['n'] = tmp[['No', 'Yes']].sum(axis=1)

weights = tmp['n'] / tmp['Yes']
weights = weights[weights.index!='other']

quotas = np.round(weights / weights.sum() * ANNOTATION_BUDGET).astype(int)
quotas_dict = dict(zip(weights.index, quotas))
quotas_dict


### embed mentions for duplicate removal and query-similarity based sampling

model_id = 'paraphrase-mpnet-base-v2'
embedder = SentenceTransformer(model_id, device='mps')

src_mentions = labeled_mentions.mention.drop_duplicates().tolist()
tgt_mentions = unlabeled_mentions.text.to_list()

embeddings = embedder.encode(src_mentions + tgt_mentions, batch_size=64, normalize_embeddings=True, show_progress_bar=True)

src_embeddings = embeddings[:len(src_mentions)]
tgt_embeddings = embeddings[len(src_mentions):]

### remove mentions that are very similar to already annotated ones

sims = cosine_similarity(tgt_embeddings, src_embeddings)

idxs = sims.max(axis=1)<0.85

unlabeled_mentions = unlabeled_mentions.iloc[idxs]
tgt_embeddings = tgt_embeddings[idxs]

len(unlabeled_mentions)

### compute mentions' similarities to attribute definitions

import json5 as json

fp = os.path.join(annotations_path, 'attribute_definitions.json')
with open(fp) as f:
    attribute_defs = json.load(f)

adefs_embeddings = embedder.encode(list(attribute_defs.values()))

### score unlabeled mentions' similarity to attribute definitions

scored = cosine_similarity(tgt_embeddings, adefs_embeddings)
scored.shape # shape (# mention, # attributes)

### select most likely mentions for each attribute dimension

idxs = {}

quotas = [quotas_dict[a] for a in attribute_defs.keys()]
for i, (a, n) in enumerate(zip(attribute_defs.keys(), quotas)):
    these = np.argsort(scored[:,i])[::-1][:n]
    idxs[a] = unlabeled_mentions.iloc[these]

sampled = pd.concat(idxs).reset_index(level=0, names=['attribute'])
len(sampled)

for a in attribute_defs.keys():
    print(a, sampled.loc[sampled.attribute==a, 'text'].iloc[:5].to_list(), sep = ':\t')

# NOTE: because some mentions are in the most-similar cases of multiple dimensions (as expected due to multilabel logic), fewer unique mentions than annotation budget
print(sampled.groupby('text').ngroups)
# sampled.groupby('text').size().sort_values(ascending=False)

## Write to disk

### aggregate at mention level

cols = ['sentence_id', 'mention_id', 'sentence_text','text']
out = sampled.groupby(cols).agg({'attribute': list}).reset_index()
out = out.rename(columns={'text': 'mention', 'sentence_text': 'text', 'attribute': 'attribute_candidates'})

# tmp = sampled.groupby(unlabeled_mentions.columns.tolist()).agg({'attribute': (list, 'count')}).reset_index().sort_values('mention_id')
# tmp[tmp.iloc[:, -1] == 3]

### add prev and following sentences

fp = os.path.join(data_path, 'manifestos', 'all_manifesto_sentences_translated.tsv')
sentences_df = pd.read_csv(fp, sep='\t')

sentences_df.rename(columns={'text_mt_m2m_100_1.2b': 'text_en'}, inplace=True)
sentences_df['manifesto_id'] = sentences_df.sentence_id.str.split('-', expand=True)[0]

# for each line, within manifesto, get the two texts occuring before and after the line in separate columns 
sentences_df['prev_texts'] = sentences_df.groupby('manifesto_id')['text_en'].shift([2, 1], fill_value='').values.tolist()
sentences_df['prev_texts'] = sentences_df['prev_texts'].apply(lambda x: [t for t in x if t != ''])
sentences_df['next_texts'] = sentences_df.groupby('manifesto_id')['text_en'].shift([-1, -2], fill_value='').values.tolist()
sentences_df['next_texts'] = sentences_df['next_texts'].apply(lambda x: [t for t in x if t != ''])

# merge with sampled sentenes
out = pd.merge(out, sentences_df[['sentence_id', 'prev_texts', 'next_texts']], on='sentence_id', how='left')

### write to disk

dest = os.path.join(annotations_path, 'social-group-mention-categorization-round03')
os.makedirs(dest, exist_ok=True)

fp = os.path.join(dest, 'sample.tsv') 
if not os.path.exists(fp):
    out.to_csv(fp, sep='\t', index=False)
