==> code/mention-classification/round4/04-combine_parsed_datasets.ipynb <==
from types import SimpleNamespace

args = SimpleNamespace()

args.data_path = '../../../data/annotations/group_mention_categorization/social-group-mention-categorization-round03'

args.text_data_file = '../../../data/labeled/manifesto_sentences_predicted_group_mentions_spans.tsv'
args.text_col = 'sentence_text'
args.text_id_col = 'sentence_id'
args.mention_col = 'text'
args.mention_id_col = 'span_nr'
args.mention_id_format = '{text_id}-{mention_id}'

args.attributes_file = '../../../data/annotations/group_mention_categorization/group_attributes_v2.yaml'

import os
import yaml
import pandas as pd

with open(args.attributes_file, 'r') as f:
    ontology = yaml.load(f, Loader=yaml.FullLoader)

attributes = pd.DataFrame([
    {
        'q_id': a,
        'q_category': i,
        'label': v,
    }
    for a, d in ontology['social_group'].items()
    for i, v in enumerate(d['attributes'].keys(), start=1)
])
attributes.q_id = attributes.q_id.str.replace('non_', 'non-')
attributes.label = attributes.label.str.replace('<i>Other attribute</i>', 'other')

attributes.q_category = attributes.q_category.astype(str)

econ_attributes_map = {c: l for _, c, l in attributes.loc[attributes.q_id == 'economic_attributes', ['q_category', 'label']].itertuples()}
nonecon_attributes_map = {c: l for _, c, l in attributes.loc[attributes.q_id == 'non-economic_attributes', ['q_category', 'label']].itertuples()}

fp = os.path.join(args.data_path, 'parsed', 'annotations.tsv')
annotations = pd.read_csv(fp, sep='\t')

annotations.loc[annotations.q_category.isna(), 'q_category'] = -1
annotations['q_category'] = annotations.q_category.astype(int)
annotations.loc[annotations.category.isna(), 'category'] = ''

annotations.mention_id.nunique()

annotations.head()

fp = os.path.join(args.data_path, 'parsed', 'reviewed_examples.tsv')
consolidated = pd.read_csv(fp, sep='\t')

# apply expert decision to disagreement cases
# consolidated.loc[consolidated.decision_hauke.notna(), 'response'] = consolidated.loc[consolidated.decision_hauke.notna(), 'decision_hauke']
consolidated.loc[consolidated.decision_hauke.notna(), 'response'] = consolidated.loc[consolidated.decision_hauke.notna(), 'decision_hauke']
consolidated['response'] = consolidated.response.str.title()

consolidated.loc[consolidated.category.isna(), 'category'] = ''

consolidated.mention_id.nunique()

out = annotations[annotations.q_id.isin(['universal_attributes', 'economic_attributes', 'non-economic_attributes'])]

cols = out.columns[:-2].tolist()
out = out.groupby(cols)[out.columns].agg({'response': lambda x: '; '.join(x[x.notna()].unique().tolist())}).reset_index()
out['coder_disagreement'] = out.response.str.contains('; ')

out['q_vals'] = out.q_id.map({'universal_attributes': 1, 'economic_attributes': 2, 'non-economic_attributes': 3})
out.sort_values(by=['mention_id', 'q_vals', 'q_category'], inplace=True, ascending=True)
del out['q_vals']

out.mention_id.nunique()

aggreement_cases = out.groupby('mention_id').filter(lambda x: ~x.coder_disagreement.any())
del aggreement_cases['coder_disagreement']

aggreement_cases.mention_id.nunique()

annotations.mention_id.nunique(), aggreement_cases.mention_id.nunique(), consolidated.mention_id.nunique()

out = pd.concat([aggreement_cases, consolidated[aggreement_cases.columns]], axis=0, ignore_index=True)

# TODO: discard this line
out = out.groupby('mention_id').filter(lambda x: ~(x.response=='??').any())

out.mention_id.nunique()

out['q_id'].value_counts()

out.loc[out['q_id']=='universal_attributes', 'response'].value_counts()
# NOTE: this is expected because of attribute query-based sampling

out.loc[out['q_id']=='economic_attributes', ['category', 'response']].value_counts(sort=False).unstack().fillna(0).astype(int)
# NOTE: this worked nicely as we have succeeded in oversampling attribute instances that were underrepresented when sampling at random

out.loc[out['q_id']=='non-economic_attributes', ['category', 'response']].value_counts(sort=False).unstack().fillna(0).astype(int)
# NOTE: this worked nicely as we have succeeded in oversampling attribute instances that were underrepresented when sampling at random

# sanity check
out.apply(lambda r: r.mention not in r.text, axis=1).sum()

# texts = pd.read_csv(args.text_data_file, sep='\t')
# texts['mention_id'] = texts.text_id.astype(str) + '-' + texts.mention_nr.astype(str)
# texts = texts[['mention_id', 'prev_texts', 'next_texts']]
texts = pd.read_csv(args.text_data_file, sep='\t')
texts['mention_id'] = texts.apply(lambda r: args.mention_id_format.format(text_id=r[args.text_id_col], mention_id=r[args.mention_id_col]), axis=1)
texts['manifesto_id'] = texts[args.text_id_col].str.split('-').str[0]
texts.rename(columns={args.mention_col: 'mention', args.text_col: 'text', 'label': 'group_type'}, inplace=True)
texts.head()

# get leading and trailing texts with shift window wdith 2
texts['prev_texts'] = texts.groupby('manifesto_id').text.shift((2, 1)).apply(lambda x: '  '.join(x.dropna().tolist()), axis=1)
texts['next_texts'] = texts.groupby('manifesto_id').text.shift((-1, -2)).apply(lambda x: '  '.join(x.dropna().tolist()), axis=1)

df = out.merge(texts[['mention_id', 'prev_texts', 'next_texts']], how='left', indicator=True)
assert all(df._merge == 'both')

df.rename(columns={'response': 'label'}, inplace=True)

cols = ['mention_id', 'text', 'mention', 'prev_texts', 'next_texts', 'q_id', 'q_category', 'category', 'label']
df = df[cols]

fp = os.path.join(args.data_path, 'parsed', 'consolidated_annotations.tsv')
df.to_csv(fp, sep='\t', index=False)
